Forms como componentes e funçao signup criada
