+++ Semana 12 - Revisão/Projeto - LabEddit/labeddit/src/APIConfig/token.js
@@ -1,0 +1,1 @@
+export const token = window.localStorage.getItem('token')
