+++ Semana 12 - Revisão/Projeto - LabEddit/labeddit/src/APIConfig/urlApi.js
@@ -1,0 +1,1 @@
+export const urlApi = 'https://us-central1-labenu-apis.cloudfunctions.net/labEddit'
